fix: updating data sources and resources for fabric cloud router
