+++ docs/data-sources/equinix_fabric_cloud_router.md
@@ -23,18 +23,20 @@ data "equinix_fabric_cloud_router" "cloud_router_data_name" {
 <!-- schema generated by tfplugindocs -->
 ## Schema
 
-### Optional
+### Required
 
 - `uuid` (String) Equinix-assigned Fabric Cloud Router identifier
 
 ### Read-Only
 
 - `account` (Set of Object) Customer account information that is associated with this Fabric Cloud Router (see [below for nested schema](#nestedatt--account))
-- `bgp_ipv4_routes_count` (Number)
-- `bgp_ipv6_routes_count` (Number)
+- `bgp_ipv4_routes_count` (Number) Access point used and maximum number of IPv4 BGP routes
+- `bgp_ipv6_routes_count` (Number) Access point used and maximum number of IPv6 BGP routes
 - `change_log` (Set of Object) Captures Fabric Cloud Router lifecycle change information (see [below for nested schema](#nestedatt--change_log))
-- `connections_count` (Number)
+- `connections_count` (Number) Number of connections associated with this Access point
 - `description` (String) Customer-provided Fabric Cloud Router description
+- `distinct_ipv4_prefixes_count` (Number) Number of distinct ipv4 routes
+- `distinct_ipv6_prefixes_count` (Number) Number of distinct ipv6 routes
 - `equinix_asn` (Number) Equinix ASN
 - `href` (String) Fabric Cloud Router URI information
 - `id` (String) The ID of this resource.
@@ -42,10 +44,10 @@ data "equinix_fabric_cloud_router" "cloud_router_data_name" {
 - `name` (String) Fabric Cloud Router name. An alpha-numeric 24 characters string which can include only hyphens and underscores
 - `notifications` (List of Object) Preferences for notifications on Fabric Cloud Router configuration or status changes (see [below for nested schema](#nestedatt--notifications))
 - `order` (Set of Object) Order information related to this Fabric Cloud Router (see [below for nested schema](#nestedatt--order))
-- `package` (Set of Object) Fabric Cloud Router package information (see [below for nested schema](#nestedatt--package))
-- `project` (Block Set) Project information (see [below for nested schema](#nestedblock--project))
+- `package` (Set of Object) Fabric Cloud Router location (see [below for nested schema](#nestedatt--package))
+- `project` (Set of Object) Fabric Cloud Router project (see [below for nested schema](#nestedatt--project))
 - `state` (String) Fabric Cloud Router overall state
-- `type` (String) Defines the Fabric Cloud Router type like XF_GATEWAY
+- `type` (String) Defines the FCR type like XF_ROUTER
 
 <a id="nestedatt--account"></a>
 ### Nested Schema for `account`
@@ -85,7 +87,6 @@ Read-Only:
 
 Read-Only:
 
-- `href` (String)
 - `ibx` (String)
 - `metro_code` (String)
 - `metro_name` (String)
@@ -121,12 +122,12 @@ Read-Only:
 - `code` (String)
 
 
-<a id="nestedblock--project"></a>
+<a id="nestedatt--project"></a>
 ### Nested Schema for `project`
 
 Read-Only:
 
-- `href` (String) Unique Resource URL
-- `project_id` (String) Project Id
+- `href` (String)
+- `project_id` (String)
