+++ docs/resources/equinix_fabric_cloud_router.md
@@ -45,27 +45,47 @@ resource "equinix_fabric_cloud_router" "new_cloud_router"{
 
 ### Required
 
+- `account` (Block Set, Min: 1, Max: 1) Customer account information that is associated with this Fabric Cloud Router (see [below for nested schema](#nestedblock--account))
 - `location` (Block Set, Min: 1, Max: 1) Fabric Cloud Router location (see [below for nested schema](#nestedblock--location))
 - `name` (String) Fabric Cloud Router name. An alpha-numeric 24 characters string which can include only hyphens and underscores
 - `notifications` (Block List, Min: 1) Preferences for notifications on Fabric Cloud Router configuration or status changes (see [below for nested schema](#nestedblock--notifications))
-- `package` (Block Set, Min: 1, Max: 1) Fabric Cloud Router package (see [below for nested schema](#nestedblock--package))
-- `type` (String) Defines the FCR type like XF_GATEWAY
+- `package` (Block Set, Min: 1, Max: 1) Fabric Cloud Router location (see [below for nested schema](#nestedblock--package))
+- `project` (Block Set, Min: 1, Max: 1) Fabric Cloud Router project (see [below for nested schema](#nestedblock--project))
+- `type` (String) Defines the FCR type like XF_ROUTER
 
 ### Optional
 
-- `account` (Block Set, Max: 1) Customer account information that is associated with this Fabric Cloud Router (see [below for nested schema](#nestedblock--account))
 - `description` (String) Customer-provided Fabric Cloud Router description
-- `order` (Block Set, Max: 1) Order information related to this Fabric Cloud Router (see [below for nested schema](#nestedblock--order))
-- `project` (Block Set) Fabric Cloud Router project (see [below for nested schema](#nestedblock--project))
+- `href` (String) Fabric Cloud Router URI information
 - `timeouts` (Block, Optional) (see [below for nested schema](#nestedblock--timeouts))
+- `uuid` (String) Equinix-assigned Fabric Cloud Router identifier
 
 ### Read-Only
 
+- `bgp_ipv4_routes_count` (Number) Access point used and maximum number of IPv4 BGP routes
+- `bgp_ipv6_routes_count` (Number) Access point used and maximum number of IPv6 BGP routes
 - `change_log` (Set of Object) Captures Fabric Cloud Router lifecycle change information (see [below for nested schema](#nestedatt--change_log))
+- `connections_count` (Number) Number of connections associated with this Access point
+- `distinct_ipv4_prefixes_count` (Number) Number of distinct ipv4 routes
+- `distinct_ipv6_prefixes_count` (Number) Number of distinct ipv6 routes
 - `equinix_asn` (Number) Equinix ASN
-- `href` (String) Fabric Cloud Router URI information
 - `id` (String) The ID of this resource.
+- `order` (Set of Object) Order information related to this Fabric Cloud Router (see [below for nested schema](#nestedatt--order))
 - `state` (String) Fabric Cloud Router overall state
+
+<a id="nestedblock--account"></a>
+### Nested Schema for `account`
+
+Read-Only:
+
+- `account_name` (String) Account Name
+- `account_number` (Number) Account Number
+- `global_cust_id` (String) Global Customer organization identifier
+- `global_org_id` (String) Global organization identifier
+- `global_organization_name` (String) Global organization name
+- `org_id` (Number) Customer organization identifier
+- `organization_name` (String) Customer organization name
+
 
 <a id="nestedblock--location"></a>
 ### Nested Schema for `location`
@@ -99,35 +119,16 @@ Required:
 - `code` (String) Fabric Cloud Router package code
 
 
-<a id="nestedblock--account"></a>
-### Nested Schema for `account`
-
-Optional:
-
-- `account_number` (Number) Account Number
-
-
-<a id="nestedblock--order"></a>
-### Nested Schema for `order`
-
-Optional:
-
-- `billing_tier` (String) Billing tier for connection bandwidth
-- `purchase_order_number` (String) Purchase order number
-
-Read-Only:
-
-- `order_id` (String) Order Identification
-- `order_number` (String) Order Reference Number
-
-
 <a id="nestedblock--project"></a>
 ### Nested Schema for `project`
 
 Optional:
 
-- `href` (String) Unique Resource URL
 - `project_id` (String) Project Id
+
+Read-Only:
+
+- `href` (String) Unique Resource URL
 
 
 <a id="nestedblock--timeouts"></a>
@@ -159,4 +160,14 @@ Read-Only:
 - `updated_by_full_name` (String)
 - `updated_date_time` (String)
 
+
+<a id="nestedatt--order"></a>
+### Nested Schema for `order`
+
+Read-Only:
+
+- `billing_tier` (String)
+- `order_id` (String)
+- `order_number` (String)
+- `purchase_order_number` (String)
