--- equinix/fabric_cloud_router_schema.go
@@ -1,168 +1,0 @@
-package equinix
-
-import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-)
-
-func createPackageSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"code": {
-			Type:        schema.TypeString,
-			Required:    true,
-			Description: "Fabric Cloud Router package code",
-		},
-	}
-}
-
-var createPackageRes = &schema.Resource{
-	Schema: createPackageSch(),
-}
-
-var createCloudRouterAccountRes = &schema.Resource{
-	Schema: createCloudRouterAccountSch(),
-}
-
-func createCloudRouterAccountSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"account_number": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Optional:    true,
-			Description: "Account Number",
-		},
-	}
-}
-
-var createCloudRouterProjectSchRes = &schema.Resource{
-	Schema: createCloudRouterProjectSch(),
-}
-
-func createCloudRouterProjectSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"project_id": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Optional:    true,
-			Description: "Project Id",
-		},
-		"href": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Computed:    true,
-			Description: "Unique Resource URL",
-		},
-	}
-}
-
-func createCloudRouterResourceSchema() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Fabric Cloud Router URI information",
-		},
-		"name": {
-			Type:        schema.TypeString,
-			Required:    true,
-			Description: "Fabric Cloud Router name. An alpha-numeric 24 characters string which can include only hyphens and underscores",
-		},
-		"description": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Customer-provided Fabric Cloud Router description",
-		},
-		"state": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Fabric Cloud Router overall state",
-		},
-		"equinix_asn": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Equinix ASN",
-		},
-		"package": {
-			Type:        schema.TypeSet,
-			Required:    true,
-			Description: "Fabric Cloud Router location",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createPackageSch(),
-			},
-		},
-		"change_log": {
-			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Captures Fabric Cloud Router lifecycle change information",
-			Elem: &schema.Resource{
-				Schema: createChangeLogSch(),
-			},
-		},
-		"type": {
-			Type:         schema.TypeString,
-			Required:     true,
-			ValidateFunc: validation.StringInSlice([]string{"XF_ROUTER"}, true),
-			Description:  "Defines the FCR type like XF_ROUTER",
-		},
-		"location": {
-			Type:        schema.TypeSet,
-			Required:    true,
-			Description: "Fabric Cloud Router location",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createLocationSch(),
-			},
-		},
-		"project": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Description: "Fabric Cloud Router project",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createCloudRouterProjectSch(),
-			},
-		},
-		"account": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Description: "Customer account information that is associated with this Fabric Cloud Router",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createCloudRouterAccountSch(),
-			},
-		},
-		"order": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Description: "Order information related to this Fabric Cloud Router",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createOrderSch(),
-			},
-		},
-		"notifications": {
-			Type:        schema.TypeList,
-			Required:    true,
-			Description: "Preferences for notifications on Fabric Cloud Router configuration or status changes",
-			Elem: &schema.Resource{
-				Schema: createNotificationSch(),
-			},
-		},
-		"bgp_ipv4_routes_count": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Access point used and maximum number of IPv4 BGP routes",
-		},
-		"bgp_ipv6_routes_count": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Access point used and maximum number of IPv6 BGP routes",
-		},
-		"connections_count": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Number of connections associated with this Access point",
-		},
-	}
-}
